requirements.txt
numpy==1.26.4
scipy==1.13.1
pandas_ta==0.3.14b0
stocktrends==0.1.5
mplfinance==0.12.10b0
ccxt==4.3.38